requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2
astroid==2.15.8 ; python_full_version >= '3.7.2'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.16.0
charset-normalizer==3.3.0 ; python_full_version >= '3.7.0'
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
dill==0.3.7 ; python_version < '3.11'
dj-database-url==2.1.0
django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
exceptiongroup==1.1.3
gunicorn==21.2.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0 ; python_full_version >= '3.8.0'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.19.1 ; python_version >= '3.8'
jsonschema-specifications==2023.7.1 ; python_version >= '3.8'
lazy-object-proxy==1.9.0 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mysql-connector-python==8.1.0
openapi-codec==1.3.2
packaging==23.2
platformdirs==3.11.0 ; python_version >= '3.7'
pluggy==1.3.0
protobuf==4.21.12
psycopg2==2.9.8
pycparser==2.21
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2 ; python_version >= '3.7' and python_version < '4.0'
pymysql==1.1.0
pytest==7.4.2
pytest-django==4.5.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.30.2 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
rpds-py==0.10.4 ; python_version >= '3.8'
simplejson==3.19.2 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1 ; python_version >= '3.7'
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6 ; python_version >= '3.7'
whitenoise==6.5.0
wrapt==1.15.0 ; python_version < '3.11'
